refactor: move all static functions to a new 'cc' package
Clearly mark out which functions relate to the test `Instance`
vs which are just true static functions.
